backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
channels>=4.0
daphne>=4.0
Pillow>=10.0
scikit-learn>=1.3
beautifulsoup4>=4.12
requests>=2.31
python-decouple>=3.8
numpy>=1.24
